requirements.txt
crewai
crewai-tools
python-dotenv
gspread
google-auth-oauthlib
pandas
streamlit
plotly
google-generativeai
langchain-google-genai
numpy
flask